backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Data Processing
pandas==2.2.1
numpy==1.26.4

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# AI & HTTP
requests==2.31.0

# CORS (for FastAPI)
python-multipart==0.0.9
fastapi[all]
# Optional: if you use other AI providers later
# openai==1.14.0


# Google API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib